Added Hibernate with Criteria

+++ maven/computer-database/src/test/java/org/excilys/computer_database/dao/DaoTest.java
@@ -138,7 +138,7 @@ public class DaoTest {
     int goodId = 1;
     Optional<Computer> computer = computerService.getComputerDetails(goodId);
     assertTrue(computer.isPresent());
-    assertEquals(goodId, computer.get().getId());
+    assertEquals(goodId, (int) computer.get().getId());
   }
 
   /**
